requirements.txt
absl-py==0.14.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
csbdeep==0.6.3
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
glob2==0.7
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==2.10.0
idna==3.2
imageio==2.9.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1598883821407/work
ipython_genutils==0.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1610146791023/work
joblib==1.0.1
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5
nmco @ git+git://github.com/GVS-Lab/chrometrics.git@786b7931e1e215806871aa0fc5cb4fda3a5808b1
numba==0.54.0
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.18.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pynndescent==0.5.4
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
pyzmq==19.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.11.2
six==1.16.0
stardist==0.6.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.2.26
tornado==6.0.4
tqdm==4.58.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
umap-learn==0.5.1
urllib3==1.26.7
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==2.0.1
wrapt==1.12.1
